final update
a bunch of things are missing but unfortunately I did this project by myself without any member of my supposed "team"

+++ DA_ProjetoFinal/Controllers/MenuController.cs
@@ -169,6 +169,11 @@ namespace DA_ProjetoFinal
                 using (var context = new CantinaContext())
                 {
                     Menu menu = context.Menus.Find(id);
+                    if (menu == null)
+                    {
+                        return false;
+                    }
+
                     context.Menus.Remove(menu);
                     context.SaveChanges();
                 }
